pyproject.toml
[tool.poetry]
name = "llm-mcp-server"
version = "0.1.0"
description = "A FastAPI-based LLM MCP server."
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = ">=3.8,<4.0"
fastapi = "*"
uvicorn = {extras = ["standard"], version = "*"}
pydantic = "*"
python-dotenv = "*"
openai = "*"

[tool.poetry.dev-dependencies]
pytest = "*"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
